Adding files and updating old ones + organization
Adding new files to separate the script from index page and ticket page, so, a new folder is being added together. Also, it was created unique files to the generated ticket in another website, but continues to be a difficult to build the ticket background in a responsive ratio. The JavaScript needed to be efforted and is working fine now.